+++ state/CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.2] - 2024-03-20
+
+### Fixed
+
+- Fixed `customInitialStateProvider` providing for `StateComponentDefinition`
+- Added `updateState(newState: S)` to `StateViewModel` for manual state updates
+
+## [0.3.1] - 2024-03-10
+
 ### Added
 
 - Full integration with Compass Navigation System